Plotly or ggplot fold change plots

Plotly or ggplot fold change plots

Usage

fcPlot(
  object,
  x1var,
  x2var,
  x1Values = NULL,
  x2Values = NULL,
  pCutoff = 0.01,
  labels = c(),
  useAdjusted = FALSE,
  plotCutoff = 1,
  graphics = "ggplot",
  fontSize = 12,
  labelFontSize = 4,
  colours = c("grey", "goldenrod1", "red", "blue"),
  verbose = FALSE,
  ...
)

Arguments

object

A glmmSeq object created by glmmSeq::glmmSeq().

x1var

The name of the first (inner) x parameter

x2var

The name of the second (outer) x parameter

x1Values

Timepoints or categories in x1var used to calculate fold change. If NULL the first two levels in x1var are used.

x2Values

Categories in x2var to be compared on x and y axis.

pCutoff

The significance cut-off for colour-coding (default = 0.01)

labels

Row names or indices to label on plot

useAdjusted

whether to use adjusted p-values (must have q-values in object). Default = FALSE

plotCutoff

Which probes to include on plot by significance cut-off (default = 1, for all markers)

graphics

Graphics system to use: "ggplot" or "plotly"

fontSize

Font size

labelFontSize

Font size for labels

colours

Vector of colours to use for significance groups

verbose

Whether to print statistics

...

Other parameters to pass to plotly or ggplot

Value

Returns a plot for fold change between x1Values in one x2Value subset on x axis and fold change in the other x2Value on the y axis.

Examples

data(PEAC_minimal_load)

disp <- apply(tpm, 1, function(x) {
(var(x, na.rm = TRUE)-mean(x, na.rm = TRUE))/(mean(x, na.rm = TRUE)**2)
})

glmmFit <- glmmSeq(~ Timepoint * EULAR_6m + (1 | PATID),
                     countdata = tpm[1:5, ],
                     metadata = metadata,
                     dispersion = disp,
                     verbose = FALSE)

fcPlot(object = glmmFit,
      x1var = "Timepoint",
      x2var = "EULAR_6m",
      x2Values = c("Good", "Non-response"),
      pCutoff = 0.05,
      useAdjusted = FALSE,
      plotCutoff = 1,
      graphics = "plotly")