Fit a glmer model for an individual gene

Fit a glmer model for an individual gene

glmerApply(
  geneList,
  fullFormula,
  data,
  control,
  modelData,
  designMatrix,
  offset,
  ...
)

Arguments

geneList
List with gene expression and dispersion

fullFormula
the model formula. For more information of formula structure see lme4::glmer()

data
The sample data or metadata.

control
the glmer control (default = glmerControl(optimizer = "bobyqa")). For more information see lme4::glmerControl().

modelData
Expanded design matrix

designMatrix
The design matrix

offset
this can be used to specify an a priori known component to be included in the linear predictor during fitting. For more information see lme4::glmer().

...
Other parameters to pass to lme4::glmer()

Value

Returns a GlmmSeq object with results for gene-wise general linear mixed models