Glmm for sequencing results of a single gene

Glmm for sequencing results of a single gene

glmmGene(
  modelFormula,
  countdata,
  gene,
  metadata,
  id,
  dispersion,
  sizeFactors = NULL,
  reducedFormula = "",
  modelData = NULL,
  control = glmerControl(optimizer = "bobyqa"),
  glmerFamily = NULL,
  zeroCount = 0.125,
  removeDuplicatedMeasures = FALSE,
  removeSingles = FALSE,
  verbose = FALSE,
  ...
)

Arguments

modelFormula
the model formula. For more information of formula structure see lme4::glmer().

countdata
the sequencing data

gene
the row name in countdata to be used

metadata
a data frame of sample information

id
Column name in metadata which contains the sample IDs to be used in pairing

dispersion
a numeric for the gene dispersion

sizeFactors
size factors (default=NULL). If provided the glmer offset is set to log(sizeFactors).For more information see lme4::glmer()

reducedFormula
Reduced design formula (default="")

modelData
something something

control
the glmer control (default=glmerControl(optimizer="bobyqa")). For more information see lme4::glmerControl().

glmerFamily
The GLM family, see stats::glm and stats::family. If NULL MASS::negative.binomial is used.

zeroCount
numerical value to offset zeroes for the purpose of log (default=0.125)

removeDuplicatedMeasures
whether to remove duplicated conditions/repeated measurements for a given time point (default=FALSE).

removeSingles
whether to remove individuals with only one measurement (default=FALSE)

verbose
Logical whether to display messaging (default=FALSE)

...
Other parameters to pass to lme4::glmer().

Value

Returns the fit for the general linear mixed model of a single gene

Examples

data(PEAC_minimal_load)
disp <- apply(tpm, 1, function(x) {
(var(x, na.rm=TRUE)-mean(x, na.rm=TRUE))/(mean(x, na.rm=TRUE)**2)
})
MS4A1fit <- glmmGene(~ Timepoint * EULAR_6m + (1 | PATID),
                      gene = "MS4A1",
                      id = "PATID",
                      countdata = tpm,
                      metadata = metadata,
                      dispersion = disp["MS4A1"],
                      verbose=FALSE)

MS4A1fit

#> Generalized linear mixed model fit by maximum likelihood (Laplace
#>   Approximation) [glmerMod]
#>  Family: Negative Binomial(0.2865)  ( log )
#> Formula: count ~ Timepoint + EULAR_6m + (1 | PATID) + Timepoint:EULAR_6m
#>    Data: data
#>  Offset: offset
#>       AIC       BIC    logLik  deviance  df.resid 
#>  898.2008  920.7630 -441.1004  882.2008       116 
#> Random effects:
#>  Groups Name        Std.Dev.
#>  PATID  (Intercept) 0.2038  
#> Number of obs: 124, groups:  PATID, 82
#> Fixed Effects:
#>                          (Intercept)                             Timepoint  
#>                              2.69088                               0.03668  
#>           EULAR_6mModerate responder                 EULAR_6mNon responder  
#>                              0.22723                              -1.69489  
#> Timepoint:EULAR_6mModerate responder       Timepoint:EULAR_6mNon responder  
#>                             -0.20100                               0.17747