Glmm Sequencing qvalues

Add qvalue columns to the glmmSeq dataframe

glmmQvals(result, cutoff = 0.05, pi0 = NULL, verbose = TRUE)

Arguments

cutoff
Prints a table showing the number of probes considered significant by the pvalue cut-off (default=0.05)

pi0
It is recommended not to input an estimate of pi0. Experienced users can use their own methodology to estimate the proportion of true nulls or set it equal to 1 for the BH procedure (default = NULL).

verbose
Logical whether to print the number of significant probes (default=TRUE)

resultA
glmmSeq object created by glmmSeq::glmmSeq().

Value

Returns a GlmmSeq object with results for gene-wise general linear mixed models with adjusted p-values using the qvalue function

Examples

data(PEAC_minimal_load)

disp <- apply(tpm, 1, function(x){ 
(var(x, na.rm=TRUE)-mean(x, na.rm=TRUE))/(mean(x, na.rm=TRUE)**2) 
})

MS4A1glmm <- glmmSeq(~ Timepoint * EULAR_6m + (1 | PATID),
                     id = 'PATID',
                     countdata = tpm[1:10, ],
                     metadata = metadata,
                     dispersion = disp[1:10],
                     verbose=FALSE)

#> Warning: PAT359 has multiple entries for identical Timepoint and EULAR_6m. These will be removed.

MS4A1glmm <- glmmQvals(MS4A1glmm, pi0=1)                    

#> 
#> q_Timepoint
#> -----------
#> Not Significant 
#>               9 
#> 
#> q_EULAR_6m
#> ----------
#> Not Significant 
#>               9 
#> 
#> q_Timepoint:EULAR_6m
#> --------------------
#> Not Significant 
#>               9