GLMM with negative binomial distribution for sequencing count data

Fits many generalised linear mixed effects models (GLMM) with negative binomial distribution for analysis of overdispersed count data with random effects. Designed for longitudinal analysis of RNA-Sequencing count data. Wald type 2 Chi-squared test is used to calculate p-values.

Usage

glmmSeq(
  modelFormula,
  countdata,
  metadata,
  id = NULL,
  dispersion,
  sizeFactors = NULL,
  reducedFormula = "",
  modelData = NULL,
  designMatrix = NULL,
  control = glmerControl(optimizer = "bobyqa"),
  cores = 1,
  removeSingles = FALSE,
  zeroCount = 0.125,
  verbose = TRUE,
  returnList = FALSE,
  progress = FALSE,
  ...
)

Arguments

modelFormula

the model formula. This must be of the form "~ ..." where the structure is assumed to be "counts ~ ...". The formula must include a random effects term. For more information on formula structure for random effects see lme4::glmer()

countdata

the sequencing count data matrix with genes in rows and samples in columns

metadata

a dataframe of sample information with variables in columns and samples in rows

id

Optional. Used to specify the column in metadata which contains the sample IDs to be used in repeated samples for random effects. If not specified, the function defaults to using the variable after the "|" in the random effects term in the formula.

dispersion

a numeric vector of gene dispersion

sizeFactors

size factors (default = NULL). If provided the glmer offset is set to log(sizeFactors). For more information see`` lme4::glmer()

reducedFormula

Reduced design formula (default = "")

modelData

Optional dataframe. Default is generated by call to expand.grid using levels of variables in the formula. Used to calculate model predictions (estimated means & 95% CI) for plotting via modelPlot. It can therefore be used to add/remove points in modelPlot.

designMatrix

custom design matrix

control

the glmer optimizer control (default = glmerControl(optimizer = "bobyqa")). See lme4::glmerControl().

cores

number of cores to use. Default = 1.

removeSingles

whether to remove individuals without repeated measures (default = FALSE)

zeroCount

numerical value to offset zeroes for the purpose of log (default = 0.125)

verbose

Logical whether to display messaging (default = TRUE)

returnList

Logical whether to return results as a list or glmmSeq object (default = FALSE). Useful for debugging.

progress

Logical whether to display a progress bar

...

Other parameters to pass to lme4::glmer()

Value

Returns an S4 class GlmmSeq object with results for gene-wise general linear mixed models. A list of results is returned if returnList

is TRUE which is useful for debugging.

Details

This function is a wrapper for lme4::glmer(). Wald type 2 Chi-squared test is calculated as per car::Anova() optimised for speed. Parallelisation is provided using parallel::mclapply on Unix/Mac or parallel::parLapply on PC.

Examples

data(PEAC_minimal_load)
disp <- apply(tpm, 1, function(x) {
(var(x, na.rm = TRUE)-mean(x, na.rm = TRUE))/(mean(x, na.rm = TRUE)**2)
})
MS4A1glmm <- glmmSeq(~ Timepoint * EULAR_6m + (1 | PATID),
                     countdata = tpm[1:2, ],
                     metadata = metadata,
                     dispersion = disp,
                     verbose = FALSE)
names(attributes(MS4A1glmm))
#>  [1] "info"           "formula"        "stats"          "predict"       
#>  [5] "reducedFormula" "countdata"      "metadata"       "modelData"     
#>  [9] "optInfo"        "errors"         "vars"           "class"